requirements.txt
# Core dependencies
numpy>=1.24.0
pandas>=2.0.0
PyYAML>=6.0.0
networkx>=3.0.0
python-dateutil>=2.9.0
pytz>=2025.2
tzdata>=2025.2

# Document processing
PyMuPDF>=1.25.5  # for PDF processing
python-docx>=0.8.11  # for DOCX processing
lxml>=5.4.0  # required by python-docx
Pillow>=10.3.0,<11.0.0  # for image processing, compatible with Streamlit
pdfminer.six>=20221105  # for PDF text extraction
langchain>=0.1.0

# Machine Learning & NLP
torch>=2.0.0
transformers>=4.30.0
sentence-transformers>=2.2.0
faiss-cpu>=1.7.4
scikit-learn>=1.0.0
scipy>=1.15.3
joblib>=1.5.0
threadpoolctl>=3.6.0

# Hugging Face ecosystem
huggingface-hub>=0.31.1
tokenizers>=0.21.1
safetensors>=0.5.3
hf-xet>=1.1.0

# LLM APIs
openai>=1.0.0
anthropic>=0.8.0

# HTTP and API clients
requests>=2.32.3
httpx>=0.28.1
httpcore>=1.0.9
h11>=0.16.0
certifi>=2025.4.26
urllib3>=2.4.0
charset-normalizer>=3.4.2
idna>=3.10

# Data validation and serialization
pydantic>=2.0.0
pydantic-core>=2.33.2
typing-extensions>=4.13.2
typing-inspection>=0.4.0
annotated-types>=0.7.0

# Visualization
matplotlib>=3.7.0
pyvis>=0.3.2

# Utilities
tqdm>=4.65.0
tenacity>=8.2.0
python-dotenv>=1.1.0  # for environment variables
filelock>=3.18.0
packaging>=16.8,<25.0  # compatible with streamlit
regex>=2024.11.6
six>=1.17.0
sniffio>=1.3.1
sympy>=1.14.0
mpmath>=1.3.0
distro>=1.9.0
anyio>=4.9.0
exceptiongroup>=1.2.2
jiter>=0.9.0
fsspec>=2023.1.0,<2024.6.0  # compatible with datasets

# Development tools
jinja2>=3.1.6
MarkupSafe>=3.0.2

# Streamlit
streamlit>=1.32,<1.36  # avoids conflict with tenacity and pillow

# Testing
pytest>=7.0.0

# New additions by KG-Builder
sentence-transformers>=2.2.0 # For embeddings
scikit-learn>=1.0.0 # For clustering
python-slugify>=8.0.0  # for generating slugs from text (used in kg_builder and kg_extractor)
neo4j>=5.0.0 # For Neo4j database interaction

# Additional dependencies
accelerate>=0.20.0
bitsandbytes>=0.41.0
black>=22.0.0
isort>=5.0.0
flake8>=4.0.0
mypy>=0.900.0
